Getting presigned AWS Polly URLs now works. Additions to Vocab.
