+++ config/Vocab.conf
@@ -1112,6 +1112,7 @@ fertugs-afmæli hk
 # fimmtugs-, sextugs- og sjötugsafmæli eru í BÍN
 Digranes-kirkja kvk örn
 hugbúnaðar-lausn kvk
+Múmín-álfur kk
 
 Seltjarnarnesbær	Seltjarnarnesbær		kk	örn	NFET
 Seltjarnarnesbær	Seltjarnarnesbæ 		kk	örn	ÞFET
